refactor: move employee module into dedicated area
