+++ src/cungLaoDong/Areas/Employees/Models/EmployeeFormModel.cs
@@ -2,7 +2,7 @@ using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace cungLaoDong.Models;
+namespace cungLaoDong.Areas.Employees.Models;
 
 [Table(name: "cung_nguoilaodong")]
 public class EmployeeFormModel
