+++ src/cungLaoDong/Data/ApplicationDbContext.cs
@@ -1,4 +1,5 @@
 ﻿using cungLaoDong.Models;
+using cungLaoDong.Areas.Employees.Models;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
