+++ src/cungLaoDong/Views/Shared/_Layout.cshtml
@@ -46,8 +46,8 @@
             </a>
             <div id="employeeManagement" class="collapse" data-bs-parent="#accordionSidebar">
                 <div class="bg-white py-2 collapse-inner rounded">
-                    <a class="collapse-item" asp-controller="Employees" asp-action="Index">Tra cứu thông tin</a>
-                    <a class="collapse-item" asp-controller="Employees" asp-action="Add">Thêm mới</a>
+                    <a class="collapse-item" asp-area="Employees" asp-controller="Employees" asp-action="Index">Tra cứu thông tin</a>
+                    <a class="collapse-item" asp-area="Employees" asp-controller="Employees" asp-action="Add">Thêm mới</a>
                 </div>
             </div>
         </li>
